refactor(op-interop-filter): extract RPC eth client
